fix a wrong return type
the return type of cftal::test::check_x_real_traits<t_real, float>::eps()
should be float

+++ test/include/cftal/test/x_real.h
@@ -127,7 +127,7 @@ namespace cftal {
         template <>
         struct check_x_real_traits<t_real, float> {
             static
-            double eps() {
+            float eps() {
                 // return std::pow(2, -64);
                 return 0x1.0p-65f; //std::ldexp(1.0, -65);
             }
